tests/requirements.txt
wheel
twine
packaging==23.1
jsonpath-rw
regex
pip
PyYAML
simplejson
urllib3<=1.26.15
requests>=2.31.0
questionary
pyfiglet
termcolor
grpcio==1.58.0
mmh3
protobuf<=3.20.0
cachetools
tb-paho-mqtt-client>=2.1.2
tb-mqtt-client==1.13.5
service-identity
pyjwt==2.6.0
tb-rest-client
pyopenssl
opcua
asyncua
twisted
pymodbus==3.0.0
pyserial
pyserial-asyncio
python-can
python-dateutil
bacpypes>=0.18.0
aiohttp
puresnmp>=2.0.0
slixmpp
psutil
orjson
pybase64
